src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.9.0",
    "websocket-client>=1.8.0",
    "websockets>=15.0",
]
